Move the ArrayHelper class to the "Util" namespace
