+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Minor fixes / code refactoring:
 - Replace deprecated Doctrine functions
 - Fix conditions not working properly in functional tests
 - Move the converter name > classname resolution to a new class named "ConverterResolver"
+- Move the ArrayHelper class to the "Util" namespace
 
 ## [2.0.3] - 2020-10-05
 [2.0.3]: https://github.com/Smile-SA/gdpr-dump/compare/2.0.2...2.0.3
